[#5993] refactor: Move the JdbcAuthorizationPlugin to authorization-common module (#5994)
### What changes were proposed in this pull request?

 Move the JdbcAuthorizationPlugin to authorization-common module

### Why are the changes needed?


Fix: #5993

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?
Just refactor.
