+++ authorizations/authorization-common/src/main/java/org/apache/gravitino/authorization/jdbc/JdbcAuthorizationPlugin.java
@@ -40,7 +40,6 @@ import org.apache.gravitino.authorization.Role;
 import org.apache.gravitino.authorization.RoleChange;
 import org.apache.gravitino.authorization.SecurableObject;
 import org.apache.gravitino.authorization.User;
-import org.apache.gravitino.authorization.common.JdbcAuthorizationProperties;
 import org.apache.gravitino.connector.authorization.AuthorizationPlugin;
 import org.apache.gravitino.exceptions.AuthorizationPluginException;
 import org.apache.gravitino.meta.AuditInfo;
@@ -55,7 +54,7 @@ import org.slf4j.LoggerFactory;
  * JDBC-based authorization plugins can inherit this class and implement their own SQL statements.
  */
 @Unstable
-abstract class JdbcAuthorizationPlugin implements AuthorizationPlugin, JdbcAuthorizationSQL {
+public abstract class JdbcAuthorizationPlugin implements AuthorizationPlugin, JdbcAuthorizationSQL {
 
   private static final String GROUP_PREFIX = "GRAVITINO_GROUP_";
   private static final Logger LOG = LoggerFactory.getLogger(JdbcAuthorizationPlugin.class);
